[REFACTOR] - 검색 기능 전부를 Search.tsx 컴포넌트로 분리
